Add test that includes one dir

+++ test/io/test_filesystem.f90
@@ -25,7 +25,8 @@ contains
             new_unittest("fs_run_valid_command", fs_run_valid_command), &
             new_unittest("fs_list_dir_empty", fs_list_dir_empty), &
             new_unittest("fs_list_dir_one_file", fs_list_dir_one_file), &
-            new_unittest("fs_list_dir_two_files", fs_list_dir_two_files) &
+            new_unittest("fs_list_dir_two_files", fs_list_dir_two_files), &
+            new_unittest("fs_list_dir_one_file_one_dir", fs_list_dir_one_file_one_dir) &
             ]
     end
 
@@ -177,6 +178,44 @@ contains
         call check(error, size(files) == 2, "The directory should contain two files.")
         call check(error, char(files(1)) == filename1, "The file should be '"//filename1//"'.")
         call check(error, char(files(2)) == filename2, "The file should be '"//filename2//"'.")
+
+        call run('rm -rf '//temp_list_dir, iostat=stat)
+    end
+
+    subroutine fs_list_dir_one_file_one_dir(error)
+        type(error_type), allocatable, intent(out) :: error
+
+        integer :: stat
+
+        type(string_type), allocatable :: contents(:)
+        character(*), parameter :: filename1 = 'abc.txt'
+        character(*), parameter :: dir = 'xyz'
+
+        call run('rm -rf '//temp_list_dir, iostat=stat)
+        if (stat /= 0) then
+            call test_failed(error, "Removing directory '"//temp_list_dir//"' failed."); return
+        end if
+
+        call run('mkdir '//temp_list_dir, iostat=stat)
+        if (stat /= 0) then
+            call test_failed(error, "Creating directory '"//temp_list_dir//"' failed."); return
+        end if
+
+        call run('touch '//temp_list_dir//'/'//filename1, iostat=stat)
+        if (stat /= 0) then
+            call test_failed(error, "Creating file 1 in directory '"//temp_list_dir//"' failed."); return
+        end if
+
+        call run('mkdir '//temp_list_dir//'/'//dir, iostat=stat)
+        if (stat /= 0) then
+            call test_failed(error, "Creating dir in directory '"//temp_list_dir//"' failed."); return
+        end if
+
+        call list_dir(temp_list_dir, contents, stat)
+        call check(error, stat, "Listing the contents of an empty directory shouldn't fail.")
+        call check(error, size(contents) == 2, "The directory should contain two files.")
+        call check(error, char(contents(1)) == filename1, "The file should be '"//filename1//"'.")
+        call check(error, char(contents(2)) == dir, "The file should be '"//dir//"'.")
 
         call run('rm -rf '//temp_list_dir, iostat=stat)
     end
